further removal of odd package uploads
