+++ ledger/sandbox-common/src/test/lib/scala/platform/sandbox/UploadPackageHelper.scala
@@ -1,0 +1,66 @@
+package com.daml.platform.sandbox
+
+import com.daml.grpc.adapter.ExecutionSequencerFactory
+import com.daml.ledger.client.configuration.{
+  CommandClientConfiguration,
+  LedgerClientChannelConfiguration,
+  LedgerClientConfiguration,
+  LedgerIdRequirement,
+}
+import com.daml.ledger.client.services.admin.PackageManagementClient
+import com.daml.ledger.client.withoutledgerid.LedgerClient
+import com.daml.ledger.runner.common.Config
+import com.daml.ledger.sandbox.SandboxOnXForTest
+import com.daml.ports.Port
+import com.google.protobuf
+
+import java.io.File
+import java.nio.file.Files
+import scala.concurrent.{ExecutionContext, Future}
+
+object UploadPackageHelper extends SandboxRequiringAuthorizationFuns {
+
+  def adminLedgerClient(port: Port, config: Config)(implicit
+      ec: ExecutionContext,
+      esf: ExecutionSequencerFactory,
+  ): LedgerClient = {
+    val participantConfig = config.participants(SandboxOnXForTest.ParticipantId)
+    val sslContext = participantConfig.apiServer.tls.flatMap(_.client())
+    val clientConfig = LedgerClientConfiguration(
+      applicationId = "admin-client",
+      ledgerIdRequirement = LedgerIdRequirement.none,
+      commandClient = CommandClientConfiguration.default,
+      token = Some(toHeader(adminTokenStandardJWT)),
+    )
+    LedgerClient.singleHost(
+      hostIp = "localhost",
+      port = port.value,
+      configuration = clientConfig,
+      channelConfig = LedgerClientChannelConfiguration(sslContext),
+    )(ec, esf)
+  }
+
+  private def uploadDarFiles(
+      client: PackageManagementClient,
+      files: List[File],
+  )(implicit
+      ec: ExecutionContext
+  ): Future[List[Unit]] =
+    if (files.isEmpty) Future.successful(List())
+    else
+      Future.sequence(files.map(uploadDarFile(client)))
+
+  private def uploadDarFile(client: PackageManagementClient)(file: File): Future[Unit] =
+    client.uploadDarFile(
+      protobuf.ByteString.copyFrom(Files.readAllBytes(file.toPath))
+    )
+
+  def uploadDarFiles(
+      client: LedgerClient,
+      files: List[File],
+  )(implicit
+      ec: ExecutionContext
+  ): Future[List[Unit]] =
+    uploadDarFiles(client.packageManagementClient, files)
+
+}
