+++ ledger/sandbox-common/src/test/lib/scala/platform/sandbox/services/TestCommands.scala
@@ -3,16 +3,8 @@
 
 package com.daml.platform.sandbox.services
 
-import com.daml.ledger.api.auth.client.LedgerCallCredentials
-
-import java.io.File
-import java.util
 import com.daml.ledger.api.domain
 import com.daml.ledger.api.testing.utils.{MockMessages => M}
-import com.daml.ledger.api.v1.admin.package_management_service.{
-  PackageManagementServiceGrpc,
-  UploadDarFileRequest,
-}
 import com.daml.ledger.api.v1.command_service.SubmitAndWaitRequest
 import com.daml.ledger.api.v1.command_submission_service.SubmitRequest
 import com.daml.ledger.api.v1.commands.Command.Command.{Create, Exercise}
@@ -24,13 +16,10 @@ import com.daml.lf.archive.DarReader
 import com.daml.lf.data.Ref.PackageId
 import com.daml.platform.participant.util.ValueConversions._
 import com.daml.platform.testing.TestTemplateIdentifiers
-import com.google.protobuf
-import com.google.protobuf.empty.Empty
-import io.grpc.Channel
 import scalaz.syntax.tag._
 
-import java.nio.file.Files
-import scala.concurrent.{ExecutionContext, Future}
+import java.io.File
+import java.util
 
 trait TestCommands {
 
@@ -41,31 +30,6 @@ trait TestCommands {
   protected def packageId: PackageId = DarReader.assertReadArchiveFromFile(darFile).main.pkgId
 
   protected def templateIds = new TestTemplateIdentifiers(packageId)
-
-  private def uploadDarFileRequest(file: File) =
-    new UploadDarFileRequest(protobuf.ByteString.copyFrom(Files.readAllBytes(file.toPath)))
-
-  def uploadPackageFiles(
-      packageFiles: List[File],
-      channel: Channel,
-      adminToken: String,
-  )(implicit ec: ExecutionContext): Future[Empty] =
-    if (packageFiles.nonEmpty) {
-      val packageManagementService =
-        LedgerCallCredentials.authenticatingStub(
-          PackageManagementServiceGrpc.stub(channel),
-          adminToken,
-        )
-      Future
-        .sequence(
-          packageFiles.map(file =>
-            packageManagementService.uploadDarFile(uploadDarFileRequest(file))
-          )
-        )
-        .map(_ => Empty())
-    } else {
-      Future.successful(Empty())
-    }
 
   protected def buildRequest(
       ledgerId: domain.LedgerId,
